requirements.txt
Django~=5.2.6
playwright~=1.55.0
requests~=2.32.5
tqdm~=4.67.1
python-dotenv~=1.1.1
gunicorn
psycopg2-binary

django-contrib-comments~=2.2.0
django-comments-xtd~=2.10.9
